pyproject.toml
[project]
name = "truss-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "baseten-performance-client>=0.0.11",
    "huggingface-hub>=0.36.0",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "truss>=0.11.23",
]
